Loading data into app
Here we are in the content page, we have routed ourself from home page
to here by using GetXon the onPressed button. Once we are here we wanted
to load two of json files that we have in our app. Both the files are
loaded in the initState and their loading is done via _readData function

Since loading operation is performed, this is an async function.